refactor: modulos
adicionado modulos do ddd
