+++ src/domain/checkout/repository/order-repository.interface.ts
@@ -1,0 +1,4 @@
+import { RepositoryInterface } from "../../@shared/repository/repository-interface";
+import { Order } from "../entity/order";
+
+export interface OrderRepositoryInterface extends RepositoryInterface<Order> {}
