+++ src/domain/customer/repository/customer-repository.interface.ts
@@ -1,0 +1,4 @@
+import { RepositoryInterface } from "../../@shared/repository/repository-interface";
+import { Customer } from "../entity/customer";
+
+export interface CustomerRepositoryInterface extends RepositoryInterface<Customer>{}
