--- src/domain/entity/customer.ts
@@ -1,76 +1,0 @@
-import Address from "./address";
-
-export class Customer {
-
-  private _id: string
-  private _name: string = ""
-  private _address: Address
-  private _active: boolean = false
-  private _rewardPoints: number = 0
-
-  constructor(id: string,  name: string ){
-    this._id = id;
-    this._name = name
-    this.validate()
-  }
-
-  get id(){
-    return this._id
-  }
-
-  get name(){
-    return this._name
-  }
-
-  get Address(): Address {
-    return this._address
-  }
-
-  get rewardPoints(){
-    return this._rewardPoints
-  }
-
-  set Address(address: Address) {
-    this._address = address;
-  }
-
-  validate() : boolean{
-    if(this._id.length === 0){
-      throw new Error("Id is required")
-    }
-
-    if(this._name.length === 0){
-      throw new Error("Name is required")
-    }
-
-    return true
-  }
-
-  changeName(name: string){
-    this._name = name
-    this.validate()
-  }
-
-  changeAddress(address: Address){
-    this._address = address
-  }
-
-  activate(){
-    if(this._address === undefined){
-       throw new Error('Address is mandatory to activate a customer')
-    }
-    this._active = true
-  }
-
-  deactivate(){
-    this._active = false
-  }
-
-  isActive(){
-    return this._active
-  }
-
-  addRewardPoints(point: number){
-    this._rewardPoints += point
-  }
-}
