--- src/domain/event/@shared/product/handler/send-email-when-product-is-created.handler.ts
@@ -1,8 +1,0 @@
-import EventHandlerInterface from "../../event-handler.interface";
-import ProductCreatedEvent from "../product-created.event";
-
-export default class SendEmailWhenProductIsCreatedHandler implements EventHandlerInterface<ProductCreatedEvent>{
-  handle(event: ProductCreatedEvent): void {
-      console.log(`Sending email to ${event.eventData.email}`)
-  }
-}
