+++ src/domain/product/event/handlers/send-email-when-product-is-created.handler.ts
@@ -1,0 +1,9 @@
+
+import EventHandlerInterface from "../../../@shared/event/event-handler.interface";
+import ProductCreatedEvent from "../product-created.event";
+
+export default class SendEmailWhenProductIsCreatedHandler implements EventHandlerInterface<ProductCreatedEvent>{
+  handle(event: ProductCreatedEvent): void {
+      console.log(`Sending email to ${event.eventData.email}`)
+  }
+}
