+++ src/domain/product/repository/product-repository.interface.ts
@@ -1,0 +1,4 @@
+import { RepositoryInterface } from '../../@shared/repository/repository-interface';
+import { Product } from '../entity/product';
+
+export interface ProductRepositoryInterface extends RepositoryInterface<Product>{}
