--- src/domain/repository/customer-repository.interface.ts
@@ -1,4 +1,0 @@
-import { Customer } from './../entity/customer';
-import { RepositoryInterface } from './repository-interface';
-
-export interface CustomerRepositoryInterface extends RepositoryInterface<Customer>{}
