--- src/domain/repository/order-repository.interface.ts
@@ -1,4 +1,0 @@
-import { Order } from "../entity/order";
-import { RepositoryInterface } from "./repository-interface";
-
-export interface OrderRepositoryInterface extends RepositoryInterface<Order> {}
