--- src/domain/repository/product-repository.interface.ts
@@ -1,4 +1,0 @@
-import { Product } from '../entity/product';
-import { RepositoryInterface } from './repository-interface';
-
-export interface ProductRepositoryInterface extends RepositoryInterface<Product>{}
