+++ src/infra/order/repository/sequelize/order.model.ts
@@ -1,0 +1,38 @@
+import {
+  Table,
+  Model,
+  PrimaryKey,
+  Column,
+  ForeignKey,
+  BelongsTo,
+  HasMany,
+} from "sequelize-typescript";
+import CustomerModel from "../../../customer/repository/sequelize/customer.model";
+
+import OrderItemModel from "./order-item.model";
+
+
+@Table({
+  tableName: "orders",
+  timestamps: false,
+})
+export default class OrderModel extends Model {
+  @PrimaryKey
+  @Column
+  declare id: string;
+
+  //relacionamento apenas por id
+  @ForeignKey(() => CustomerModel)
+  @Column({ allowNull: false })
+  declare customer_id: string;
+
+  //isso é para trazer os dados do cliente
+  @BelongsTo(() => CustomerModel)
+  declare customer: CustomerModel;
+
+  @HasMany(() => OrderItemModel)
+  declare items: OrderItemModel[];
+
+  @Column({ allowNull: false })
+  declare total: number;
+}
