--- src/infra/repository/order.repository.spec.ts
@@ -1,269 +1,0 @@
-import { Sequelize } from "sequelize-typescript";
-import Address from "../../domain/entity/address";
-import { Customer } from "../../domain/entity/customer";
-import { Order } from "../../domain/entity/order";
-import OrderItem from "../../domain/entity/order-item";
-import { Product } from "../../domain/entity/product";
-import CustomerModel from "../db/sequelize/model/customer.model";
-import OrderItemModel from "../db/sequelize/model/order-item.model";
-import OrderModel from "../db/sequelize/model/order.model";
-import ProductModel from "../db/sequelize/model/product.model";
-import CustomerRepository from "./customer.repository";
-import OrderRepository from "./order.repository";
-import { ProductRepository } from "./product.repository";
-
-describe("Order repository test", () => {
-  let sequelize: Sequelize;
-
-  beforeEach(async () => {
-    sequelize = new Sequelize({
-      dialect: "sqlite",
-      storage: ":memory:",
-      logging: false,
-      sync: { force: true },
-    });
-
-    sequelize.addModels([
-      CustomerModel,
-      OrderModel,
-      OrderItemModel,
-      ProductModel,
-    ]);
-    await sequelize.sync();
-  });
-
-  afterEach(async () => {
-    await sequelize.close();
-  });
-
-  it("should create a new order", async () => {
-    const customerRepository = new CustomerRepository();
-    const customer = new Customer("123", "Customer 1");
-    const address = new Address("Street 1", 1, "Zipcode 1", "City 1");
-    customer.changeAddress(address);
-    await customerRepository.create(customer);
-
-    const productRepository = new ProductRepository();
-    const product = new Product("123", "Product 1", 10);
-    await productRepository.create(product);
-
-    const ordemItem = new OrderItem(
-      "1",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const order = new Order("123", customer.id, [ordemItem]);
-
-    const orderRepository = new OrderRepository();
-    await orderRepository.create(order);
-
-    const orderModel = await OrderModel.findOne({
-      where: { id: order.id },
-      include: ["items"],
-    });
-
-    expect(orderModel.toJSON()).toStrictEqual({
-      id: "123",
-      customer_id: "123",
-      total: order.total(),
-      items: [
-        {
-          id: ordemItem.id,
-          name: ordemItem.name,
-          price: ordemItem.price,
-          quantity: ordemItem.quantity,
-          order_id: "123",
-          product_id: "123",
-        },
-      ],
-    });
-  });
-
-  it('should update a order', async () => {
-    const customerRepository = new CustomerRepository();
-    const customer = new Customer("123", "Customer 1");
-    const address = new Address("Street 1", 1, "Zipcode 1", "City 1");
-    customer.changeAddress(address);
-    await customerRepository.create(customer);
-
-    const productRepository = new ProductRepository();
-    const product = new Product("123", "Product 1", 10);
-    const product2 = new Product("124", "Product 2", 5);
-    await productRepository.create(product);
-    await productRepository.create(product2);
-
-    const ordemItem1 = new OrderItem(
-      "1",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const ordemItem2 = new OrderItem(
-      "2",
-      product2.name,
-      product2.price,
-      product2.id,
-      6
-    );
-
-    const order = new Order("123", customer.id, [ordemItem1]);
-
-    const orderRepository = new OrderRepository();
-    await orderRepository.create(order);
-
-    let orderModel = await OrderModel.findOne({
-      where: { id: order.id },
-      include: ["items"],
-    });
-
-    expect(orderModel.items).toHaveLength(1)
-
-    order.addNewItem(ordemItem2)
-
-    await orderRepository.update(order)
-    orderModel = await OrderModel.findOne({
-      where: { id: order.id },
-      include: ["items"],
-    });
-    expect(orderModel.items).toHaveLength(2)
-  });
-
-  it("should find an order", async () => {
-    const customerRepository = new CustomerRepository();
-    const customer = new Customer("123", "Customer 1");
-    const address = new Address("Street 1", 1, "Zipcode 1", "City 1");
-    customer.changeAddress(address);
-    await customerRepository.create(customer);
-
-    const productRepository = new ProductRepository();
-    const product = new Product("999", "Product 1", 10);
-    await productRepository.create(product);
-
-    const ordemItem = new OrderItem(
-      "1",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const order = new Order("123", customer.id, [ordemItem]);
-
-    const orderRepository = new OrderRepository();
-    await orderRepository.create(order);
-
-    const foundOrder = await orderRepository.find(order.id)
-
-    const orderModel = await OrderModel.findOne({
-      where: { id: order.id },
-      include: ["items"],
-    });
-
-    expect(orderModel.toJSON()).toMatchObject({
-      customer_id: foundOrder.customerId,
-      id: foundOrder.id,
-      items: [
-        {
-          id: foundOrder.items[0].id,
-          name: foundOrder.items[0].name,
-          price: foundOrder.items[0].price,
-          product_id: foundOrder.items[0].productId,
-          quantity:foundOrder.items[0].quantity
-        },
-      ],
-      total: foundOrder.total(),
-    });
-  });
-
-  it("find all orders", async () => {
-    const customerRepository = new CustomerRepository();
-    const customer = new Customer("123", "Customer 1");
-    const address = new Address("Street 1", 1, "Zipcode 1", "City 1");
-    customer.changeAddress(address);
-    await customerRepository.create(customer);
-
-    const productRepository = new ProductRepository();
-    const product = new Product("123", "Product 1", 10);
-    await productRepository.create(product);
-
-    const ordemItem = new OrderItem(
-      "1",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const ordemItem2 = new OrderItem(
-      "2",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const order = new Order("123", customer.id, [ordemItem]);
-    const order2 = new Order("1234", customer.id, [ordemItem2]);
-    const orderRepository = new OrderRepository();
-  
-    await orderRepository.create(order);
-    await orderRepository.create(order2);
-
-    const orders = await orderRepository.findAll();
-
-    expect(orders.length).toBe(2);
-  });
-
-  it("should throw an error when order is not found", async () => {
-    const orderRepository = new OrderRepository();
-
-    expect(async () => {
-      await orderRepository.find("456ABC");
-    }).rejects.toThrow("Order not found");
-  });
-  
-  it("find all orders", async () => {
-    const customerRepository = new CustomerRepository();
-    const customer = new Customer("123", "Customer 1");
-    const address = new Address("Street 1", 1, "Zipcode 1", "City 1");
-    customer.changeAddress(address);
-    await customerRepository.create(customer);
-
-    const productRepository = new ProductRepository();
-    const product = new Product("123", "Product 1", 10);
-    await productRepository.create(product);
-
-    const ordemItem = new OrderItem(
-      "1",
-      product.name,
-      product.price,
-      product.id,
-      2
-    );
-
-    const ordemItem2 = new OrderItem(
-      "2",
-      product.name,
-      product.price,
-      product.id,
-      4
-    );
-
-    const order = new Order("123", "123", [ordemItem]);
-    const order2 = new Order("1234", "123", [ordemItem2]);
-    
-    const orderRepository = new OrderRepository();
-
-    await orderRepository.create(order);
-    //await orderRepository.create(order2);
-
-    const orders = await orderRepository.findAll();
-
-   // expect(orders.length).toBe(2);
-  });
-});
