--- src/infra/repository/order.repository.ts
@@ -1,99 +1,0 @@
-import { OrderRepositoryInterface } from './../../domain/repository/order-repository.interface';
-import { Order } from "../../domain/entity/order";
-import OrderItemModel from "../db/sequelize/model/order-item.model";
-import OrderModel from "../db/sequelize/model/order.model";
-import OrderItem from '../../domain/entity/order-item';
-
-export default class OrderRepository implements OrderRepositoryInterface{
-  async create(entity: Order): Promise<void> {
-    await OrderModel.create(
-      {
-        id: entity.id,
-        customer_id: entity.customerId,
-        total: entity.total(),
-        items: entity.items.map((item) => ({
-          id: item.id,
-          name: item.name,
-          price: item.price,
-          product_id: item.productId,
-          quantity: item.quantity,
-        })),
-      },
-      {
-        include: [{ model: OrderItemModel }],
-      }
-    );
-  }
-  
-  async update(entity: Order): Promise<void> {
-    const sequelize = OrderModel.sequelize;
-    await sequelize.transaction(async (t) => {
-      await OrderItemModel.destroy({
-        where: { order_id: entity.id },
-        transaction: t,
-      });
-      const items = entity.items.map((item) => ({
-        id: item.id,
-        name: item.name,
-        price: item.price,
-        product_id: item.productId,
-        quantity: item.quantity,
-        order_id: entity.id,
-      }));
-      await OrderItemModel.bulkCreate(items, { transaction: t });
-      await OrderModel.update(
-        { total: entity.total() },
-        { where: { id: entity.id }, transaction: t }
-      );
-    });
-  }
-
-  async find(id: string): Promise<Order> {
-    let order;
-    try {
-      order = await OrderModel.findOne({
-        where: {
-          id,
-        },
-        include: [{ model: OrderItemModel }],
-        rejectOnEmpty: true,
-      });
-    } catch (error) {
-      throw new Error("Order not found");
-    }
-
-    
-    const formatedItems = order.items.map(
-      (item) =>
-        new OrderItem(
-          item.id,
-          item.name,
-          item.price,
-          item.product_id,
-          item.quantity
-        )
-    );
-
-    return new Order(id, order.customer_id, formatedItems);
-  }
-
-  async findAll(): Promise<Order[]> {
-    const ordersModel = await OrderModel.findAll({
-      include: ["items"]
-    })
-  
-    return ordersModel.map(order => {
-      return new Order(
-        order.id,
-        order.customer_id,
-        order.items.map((item) => new OrderItem(
-          item.id,
-          item.name,
-          item.price,
-          item.product_id,
-          item.quantity,
-        )),
-      )
-    })
-  }
-}
